backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.3.0
pymongo>=4.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
google-genai>=1.0.0
